[Ch-#153279622] Rename verifyToken to adhere to Airbnb
